ai_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docstring_parser==0.16
et_xmlfile==2.0.0
faiss-cpu==1.10.0
fastapi==0.115.8
ffmpeg==1.4
filelock==3.17.0
fsspec==2025.2.0
git-filter-repo==2.47.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-speech==2.30.0
google-cloud-storage==2.19.0
google-cloud-translate==3.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
shapely==2.0.7
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vertexai==1.71.1
wsproto==1.2.0
